+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Redesigned the `setvars` middleware to support the `default` field.
 - The `params` field of middlewares has been changed to the `any` type.
 - Enhance validation of `upstream` and `service` in the config package.
+- Remove `skip_test` from resolver.
 
 ### **Test**